Fix types

+++ hazlab2.h
@@ -54,8 +54,8 @@ typedef struct {
 
 
 // Defines the speed of movements
-#define DELTA_ROTATION            (GLfloat)0.075
-#define DELTA_TRANSLATION_DIVISOR 10
+#define DELTA_ROTATION            0.075
+#define DELTA_TRANSLATION_DIVISOR 10.0
 
 
 // printf to standard error
